requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
bleach==6.1.0
borb==2.1.25
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
defusedxml==0.8.0rc2
dj-database-url==2.2.0
Django==5.1
django-analytical==3.1.0
django-ckeditor==6.7.1
django-constance==4.0.1
django-js-asset==2.2.0
django-summernote==0.8.20.0
fonttools==4.53.1
fpdf2==2.7.9
gunicorn==23.0.0
idna==3.8
lxml==5.3.0
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
pypng==0.20220715.0
python-barcode==0.15.1
python3-openid==3.2.0
pytz==2024.1
qrcode==7.4.2
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
soupsieve==2.6
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
webencodings==0.5.1
whitenoise==6.7.0
